+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [`13.6.2.dev0` (unreleased)](https://github.com/kdeldycke/extra-platforms/compare/v13.6.1...main)
+
+> [!WARNING]
+> This version is **not released yet** and is under active development.
+
 ## [`13.6.1` (2026-08-03)](https://github.com/kdeldycke/extra-platforms/compare/v13.6.0...v13.6.1)
 
 ## [`13.6.0` (2026-08-03)](https://github.com/kdeldycke/extra-platforms/compare/v13.5.3...v13.6.0)